+++ Controllers/PhoneBookAppController.cs
@@ -7,38 +7,16 @@ public class PhoneBookAppController
 {
     public void CreateContact()
     {
-        GetInput getInput = new GetInput();
+        GetInput getInput = new();
 
         var context = new PhoneBookAppDbContext();
 
         var newContact = new Contact();
 
-        Console.WriteLine("New Contact");
-
-        Console.WriteLine("\nContact Name or 0 to return to Menu:");
-        string contactName = Console.ReadLine();       
-        
-        if (contactName == "0") { CreateContact(); }
-
-        while (!Validation.IsNameValid(contactName))
-        {
-            Console.WriteLine("\nPlease input a valid Contact Name:");
-            contactName = Console.ReadLine();
-        }
-
-        Console.WriteLine("\nPhone Number or 0 to return to Menu:");
-        string contactNumber = Console.ReadLine();
-
-        if (contactNumber == "0") { CreateContact(); }
-
-        while (!Validation.IsPhoneNumberValid(contactNumber))
-        {
-            Console.WriteLine("\nPlease input a valid Contact Number:");
-            contactNumber = Console.ReadLine();
-        }
+        var (contactName, contactNumber) = getInput.GetNewContactInput(); // TODO : No tuple
 
         newContact.Name = contactName;
-        newContact.PhoneNumber = contactNumber;            
+        newContact.PhoneNumber = contactNumber;
 
         context.Contacts.Add(newContact);
         context.SaveChanges();
@@ -51,41 +29,14 @@ public class PhoneBookAppController
 
     public void DeleteContact()
     {
-        GetInput getInput = new GetInput();
-        DisplayTable displayTable = new DisplayTable();
+        GetInput getInput = new();
+        DisplayTable displayTable = new();
 
         displayTable.DisplayContacts();
 
-        var context = new PhoneBookAppDbContext();            
+        var context = new PhoneBookAppDbContext();
 
-        Console.WriteLine("Delete Contact");
-
-        Console.WriteLine("\nId of contact to Delete or 0 to go back to Menu:");
-        string contactId = Console.ReadLine();
-
-        if (contactId == "0") { getInput.MainMenu(); }
-
-        while (!Validation.IsIdValid(contactId))
-        {
-            Console.WriteLine("\nInvalid entry, please enter the Id of contact to Delete or 0 to go back to Menu:");
-            contactId = Console.ReadLine();
-
-            if (contactId == "0") { getInput.MainMenu(); }
-        }
-
-        int id = Int32.Parse(contactId);
-        var deleteContact = context.Contacts.Find(id);
-
-        while (deleteContact == null)
-        {
-            Console.WriteLine("\nInvalid entry, please enter a valid Id or 0 to go back to Menu:");
-            contactId = Console.ReadLine();
-
-            if (contactId == "0") { getInput.MainMenu(); }
-
-            id = Int32.Parse(contactId);
-            deleteContact = context.Contacts.Find(id);
-        }
+        var deleteContact = getInput.GetDeleteContactInput();                    
 
         Console.WriteLine($"Are you sure you want to delete the entry for {deleteContact.Name}? \n\nType \"yes\" to Delete or any other Entry to Cancel");
 
@@ -109,106 +60,15 @@ public class PhoneBookAppController
     }
 
     public void EditContact()
-    {
-        bool nameUpdtaed = true;
-        bool numberUpdtated = true;
-
-        GetInput getInput = new GetInput();
-        DisplayTable displayTable = new DisplayTable();
+    {      
+        GetInput getInput = new();
+        DisplayTable displayTable = new();
 
         displayTable.DisplayContacts();
 
         var context = new PhoneBookAppDbContext();        
 
-        Console.WriteLine("\nId of contact to Edit or 0 to go back to Menu:");
-        string contactId = Console.ReadLine();
-
-        if (contactId == "0") { getInput.MainMenu(); }
-
-        while (!Validation.IsIdValid(contactId))
-        {
-            Console.WriteLine("\nInvalid entry, please enter the Id of contact to Delete or 0 to go back to Menu:");
-            contactId = Console.ReadLine();
-
-            if (contactId == "0") { getInput.MainMenu(); }
-        }
-
-        int id = Int32.Parse(contactId);
-        var editContact = context.Contacts.Find(id);
-
-        while (editContact == null)
-        {
-            Console.WriteLine("\nInvalid entry, please enter a valid Id or 0 to go back to Menu:");
-            contactId = Console.ReadLine();
-
-            if (contactId == "0") { getInput.MainMenu(); }
-
-            id = Int32.Parse(contactId);
-            editContact = context.Contacts.Find(id);            
-        }
-
-        displayTable.DisplayContact(id);
-
-        Console.WriteLine("\nEnter update to Contact Name or 0 to return to Menu (Press Enter to use current Name):");
-        string contactName = Console.ReadLine();
-
-        if (contactName == "0") { EditContact(); }
-
-        else if (contactName == "")
-        {
-            contactName = editContact.Name.ToString();
-            nameUpdtaed = false;
-        }
-
-        while (!Validation.IsNameValid(contactName))
-        {
-            Console.WriteLine("\nPlease input a valid Contact Name:");
-            contactName = Console.ReadLine();
-
-            if (contactName == "0") { EditContact(); }
-
-            else if (contactName == "")
-            {
-                contactName = editContact.Name.ToString();
-                nameUpdtaed = false;
-            }
-        }
-
-        Console.WriteLine("\nEnter update to Phone Number or 0 to return to Menu (Press Enter to use current Phone Number):");
-        string contactNumber = Console.ReadLine();
-
-        if (contactNumber == "0") { EditContact(); }
-
-        else if (contactNumber == "")
-        {
-            contactNumber = editContact.PhoneNumber.ToString();
-            numberUpdtated= false;
-        }
-
-        while (!Validation.IsPhoneNumberValid(contactNumber))
-        {
-            Console.WriteLine("\nPlease input a valid Contact Number:");
-            contactNumber = Console.ReadLine();
-
-            if (contactNumber == "0") { EditContact(); }
-
-            else if (contactNumber == "")
-            {
-                contactNumber = editContact.PhoneNumber.ToString();
-                numberUpdtated= false;
-            }
-        }
-                
-        if (nameUpdtaed == false && numberUpdtated == false)
-        {
-            Console.WriteLine("\nContact unchanged.\nPress Enter...");
-            Console.ReadLine();
-
-            getInput.MainMenu();
-        }
-
-        editContact.Name = contactName;
-        editContact.PhoneNumber = contactNumber;
+        Contact editContact = getInput.GetEditContactInput();
 
         context.Contacts.Update(editContact);
         context.SaveChanges();
@@ -221,16 +81,16 @@ public class PhoneBookAppController
 
     public void SearchContact()
     {
-        GetInput getInput = new GetInput();            
+        GetInput getInput = new();            
 
         var context = new PhoneBookAppDbContext();
 
-        Console.WriteLine("Search Contact");
+        //Console.WriteLine("Search Contact");
 
-        Console.WriteLine("\nEnter Name or Number to search for or 0 to go back to Menu:");
-        string contactSearch = Console.ReadLine();
+        //Console.WriteLine("\nEnter Name or Number to search for or 0 to go back to Menu:");
+        //string contactSearch = Console.ReadLine();
 
-        if (contactSearch == "0") { getInput.MainMenu(); }
+        //if (contactSearch == "0") { getInput.MainMenu(); }
 
         var searchContact = context.Contacts.Where(j => j.Name.Contains(contactSearch)).ToList();
         var searchNumber = context.Contacts.Where(j => j.PhoneNumber.Contains(contactSearch)).ToList();
